tests/wpt/web-platform-tests/css/css-sizing/parsing/max-width-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Intrinsic & Extrinsic Sizing Test: parsing max-width with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-sizing-3/#propdef-max-width">
<meta name="assert" content="max-width supports only the grammar 'none | <length-percentage> | min-content | max-content | fit-content(<length-percentage>)'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("max-width", "auto");
test_invalid_value("max-width", "min-content max-content");

test_invalid_value("max-width", "-10%");
test_invalid_value("max-width", "-0.5em");
</script>
</body>
</html>
